Fix various form, type, and formatting issues

+++ src/routes/(main)/avatarshop/create/+page.server.ts
@@ -9,7 +9,7 @@ import { z } from "zod"
 
 const schema = z.object({
 	name: z.string().min(3).max(50),
-	price: z.number().min(0).max(9999),
+	price: z.number().int().min(0).max(9999),
 	category: z.enum([
 		"TShirt",
 		"Shirt",
